compact enum in registers
It in impractical to have many definitions for the same data type. There
are cases where we want to access them in the same manner. For example
all gpiox.moder.Moder0 enums are the same and hal is using one type for
modes. If every register defines new enum that complicates hal.